added netblock query
